+++ CHANGELOG.md
@@ -6,6 +6,8 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [Unreleased]
+### Added
+- HMRC header component [#808](https://github.com/hmrc/assets-frontend/pull/808)
 
 ## [2.249.0] - 2017-09-18
 ### Changed